--- src/main/java/com/github/dogyungkim/yongjibus/yongjibus/daytype/service/DayTypeService.java
@@ -1,11 +1,0 @@
-package com.github.dogyungkim.yongjibus.yongjibus.daytype.service;
-
-import com.github.dogyungkim.yongjibus.yongjibus.daytype.model.DateInfo;
-
-import java.time.LocalDate;
-
-public interface DayTypeService {
-    DateInfo getDayType(LocalDate date);
-
-    void setHolidayInfo();
-}
