+++ src/main/java/com/yongjibus/YongjibusApplication.java
@@ -1,15 +1,7 @@
-package com.github.dogyungkim.yongjibus.yongjibus;
+package com.yongjibus;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
-
-import java.text.SimpleDateFormat;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Date;
 
 @EnableScheduling
 @SpringBootApplication
