+++ src/main/java/com/yongjibus/daytype/repository/InMemoryDayTypeRepository.java
@@ -1,21 +1,23 @@
-package com.github.dogyungkim.yongjibus.yongjibus.daytype.repository;
+package com.yongjibus.daytype.repository;
 
-import com.github.dogyungkim.yongjibus.yongjibus.daytype.model.DateInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
+import com.yongjibus.daytype.model.DateInfo;
+import com.yongjibus.exception.DateInfoNotFoundException;
+
 import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.stream.IntStream;
 
 @Repository
 @RequiredArgsConstructor
-public class LocalDayTypeRepository implements DayTypeRepository {
+public class InMemoryDayTypeRepository implements DayTypeRepository {
 
     // List로 HolidayInfo를 저장
-    private static final List<DateInfo> store = new ArrayList<>();
+    private static final List<DateInfo> store = new CopyOnWriteArrayList<>();
 
     /**
      * 특정 날짜를 기준으로 HolidayInfo를 검색
@@ -29,7 +31,7 @@ public class LocalDayTypeRepository implements DayTypeRepository {
                 return dateInfo;
             }
         }
-        return new DateInfo();
+        throw new DateInfoNotFoundException(date.toString());
     }
 
     /**
