+++ src/main/java/com/yongjibus/daytype/service/DayTypeService.java
@@ -1,0 +1,11 @@
+package com.yongjibus.daytype.service;
+
+import java.time.LocalDate;
+
+import com.yongjibus.daytype.model.DateInfo;
+
+public interface DayTypeService {
+    DateInfo getDayType(LocalDate date);
+
+    void setHolidayInfo();
+}
