+++ src/main/java/com/yongjibus/exception/DateInfoNotFoundException.java
@@ -1,0 +1,7 @@
+package com.yongjibus.exception;
+
+public class DateInfoNotFoundException extends RuntimeException {
+    public DateInfoNotFoundException(String message) {
+        super("DateInfo not found : " + message);
+    }
+}
